# openspec/versions/v0.0.1/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/biblioteca/programas": {
      "get": {
        "operationId": "BibliotecaController_listarProgramas",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Lista de programas académicos del usuario"
          },
          "401": {
            "description": "Token JWT ausente o inválido"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar programas a los que el usuario tiene acceso",
        "tags": [
          "biblioteca"
        ]
      }
    },
    "/api/biblioteca/programas/{id}/recursos": {
      "get": {
        "operationId": "BibliotecaController_listarRecursos",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID del programa académico",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "tipo",
            "required": false,
            "in": "query",
            "description": "Filtrar por tipo de contenido",
            "schema": {
              "enum": [
                "ENLACE",
                "DOCUMENTO",
                "VIDEO",
                "IMAGEN",
                "ARTICULO",
                "OTRO"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de recursos decorados del programa"
          },
          "401": {
            "description": "Token JWT ausente o inválido"
          },
          "403": {
            "description": "El usuario no pertenece al programa solicitado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar recursos del programa con filtro por tipo_contenido",
        "tags": [
          "biblioteca"
        ]
      },
      "post": {
        "operationId": "BibliotecaController_crearRecurso",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID del programa académico",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateResourceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Recurso creado con metadatos OG extraídos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateResourceDto"
                }
              }
            }
          },
          "400": {
            "description": "Payload inválido — se requiere url_externa o titulo"
          },
          "401": {
            "description": "Token JWT ausente o inválido"
          },
          "403": {
            "description": "El usuario no pertenece al programa"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Crear recurso con extracción automática de metadatos Open Graph",
        "tags": [
          "biblioteca"
        ]
      }
    },
    "/api/biblioteca/recursos/{id}": {
      "get": {
        "operationId": "BibliotecaController_obtenerRecurso",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obtener recurso decorado por ID",
        "tags": [
          "biblioteca"
        ]
      },
      "patch": {
        "operationId": "BibliotecaController_editarRecurso",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateResourceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "CA3: Editar recurso (solo propietario o admin del grupo asociado)",
        "tags": [
          "biblioteca"
        ]
      },
      "delete": {
        "operationId": "BibliotecaController_eliminarRecurso",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID del recurso",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Recurso eliminado correctamente"
          },
          "401": {
            "description": "Token JWT ausente o inválido"
          },
          "403": {
            "description": "Sin permiso — no es propietario ni admin del grupo"
          },
          "404": {
            "description": "Recurso no encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Eliminar recurso (solo propietario o admin del grupo asociado)",
        "tags": [
          "biblioteca"
        ]
      }
    },
    "/api/biblioteca/recursos/{id}/comentarios": {
      "post": {
        "operationId": "BibliotecaController_comentar",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID del recurso",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddCommentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Comentario agregado"
          },
          "401": {
            "description": "Token JWT ausente o inválido"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Agregar comentario a un recurso",
        "tags": [
          "biblioteca"
        ]
      }
    },
    "/api/biblioteca/recursos/{id}/valoracion": {
      "post": {
        "operationId": "BibliotecaController_valorar",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID del recurso",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RateResourceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Valoración registrada o actualizada"
          },
          "400": {
            "description": "Valor fuera del rango 1–5"
          },
          "401": {
            "description": "Token JWT ausente o inválido"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Valorar un recurso (1–5 estrellas, upsert)",
        "tags": [
          "biblioteca"
        ]
      }
    }
  },
  "info": {
    "title": "Uniconnect API",
    "description": "Documentación oficial de la API REST de Uniconnect — red social universitaria.\n\nTodos los endpoints protegidos requieren un JWT Bearer token obtenido en `POST /api/auth/callback`.\n\n**OpenAPI JSON**: `GET /docs-json`  |  **OpenAPI YAML**: `GET /docs-yaml`",
    "version": "0.0.1",
    "contact": {
      "name": "Uniconnect Team",
      "url": "",
      "email": "luma1017hl@gmail.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:8007",
      "description": "Desarrollo local"
    },
    {
      "url": "https://uniconnect-backend.fly.dev",
      "description": "Producción (Fly.io)"
    }
  ],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateResourceDto": {
        "type": "object",
        "properties": {
          "url_externa": {
            "type": "string",
            "example": "https://arxiv.org/abs/2106.09685"
          },
          "titulo": {
            "type": "string",
            "example": "Attention is All You Need"
          },
          "descripcion": {
            "type": "string"
          },
          "tipo_contenido": {
            "type": "string",
            "enum": [
              "ENLACE",
              "DOCUMENTO",
              "VIDEO",
              "IMAGEN",
              "ARTICULO",
              "OTRO"
            ],
            "default": "ENLACE"
          },
          "etiquetas": {
            "example": [
              "IA",
              "Deep Learning"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "id_group": {
            "type": "number",
            "example": 1
          }
        }
      },
      "UpdateResourceDto": {
        "type": "object",
        "properties": {
          "titulo": {
            "type": "string"
          },
          "descripcion": {
            "type": "string"
          },
          "tipo_contenido": {
            "type": "string",
            "enum": [
              "ENLACE",
              "DOCUMENTO",
              "VIDEO",
              "IMAGEN",
              "ARTICULO",
              "OTRO"
            ]
          },
          "etiquetas": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "AddCommentDto": {
        "type": "object",
        "properties": {
          "contenido": {
            "type": "string",
            "example": "Excelente recurso, muy útil para el parcial."
          }
        },
        "required": [
          "contenido"
        ]
      },
      "RateResourceDto": {
        "type": "object",
        "properties": {
          "valor": {
            "type": "number",
            "example": 4,
            "minimum": 1,
            "maximum": 5
          }
        },
        "required": [
          "valor"
        ]
      }
    }
  }
}